JAMES-3581 Refactoring ResponseSerializer and extract PushSerializer
