+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/json/ResponseSerializer.scala
@@ -19,13 +19,9 @@
 
 package org.apache.james.jmap.json
 
-import java.io.InputStream
-import java.net.URL
-
 import eu.timepit.refined.refineV
 import io.netty.handler.codec.http.HttpResponseStatus
 import org.apache.james.core.Username
-import org.apache.james.jmap.change.{TypeName, TypeState}
 import org.apache.james.jmap.core
 import org.apache.james.jmap.core.CapabilityIdentifier.CapabilityIdentifier
 import org.apache.james.jmap.core.Id.IdConstraint
@@ -35,9 +31,10 @@ import org.apache.james.jmap.core.{Account, Invocation, Session, _}
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
+import java.io.InputStream
+import java.net.URL
 import scala.collection.{Seq => LegacySeq}
 import scala.language.implicitConversions
-import scala.util.Try
 
 object ResponseSerializer {
   // CreateIds
@@ -150,106 +147,22 @@ object ResponseSerializer {
 
   private implicit val problemDetailsWrites: OWrites[ProblemDetails] = Json.writes[ProblemDetails]
 
-  private implicit val requestIdFormat: Format[RequestId] = Json.valueFormat[RequestId]
-  private implicit val webSocketRequestReads: Reads[WebSocketRequest] = {
-    case jsObject: JsObject =>
-      for {
-        requestId <- jsObject.value.get("requestId")
-          .map(requestIdJson => requestIdFormat.reads(requestIdJson).map(Some(_)))
-          .getOrElse(JsSuccess(None))
-        request <- requestObjectRead.reads(jsObject)
-      } yield {
-        WebSocketRequest(requestId, request)
-      }
-    case _ => JsError("Expecting a JsObject to represent a webSocket inbound request")
-  }
-  private implicit val typeNameReads: Reads[TypeName] = {
-    case JsString(string) => TypeName.parse(string)
-      .fold(errorMessage => JsError(errorMessage), JsSuccess(_))
-    case _ => JsError("Expecting a JsString as typeName")
-  }
-  private implicit val pushStateReads: Reads[PushState] = Json.valueReads[PushState]
-  private implicit val webSocketPushEnableReads: Reads[WebSocketPushEnable] = Json.reads[WebSocketPushEnable]
-  private implicit val webSocketInboundReads: Reads[WebSocketInboundMessage] = {
-    case json: JsObject =>
-      json.value.get("@type") match {
-        case Some(JsString("Request")) => webSocketRequestReads.reads(json)
-        case Some(JsString("WebSocketPushEnable")) => webSocketPushEnableReads.reads(json)
-        case Some(JsString("WebSocketPushDisable")) => JsSuccess(WebSocketPushDisable)
-        case Some(JsString(unknownType)) => JsError(s"Unknown @type field on a webSocket inbound message: $unknownType")
-        case Some(invalidType) => JsError(s"Invalid @type field on a webSocket inbound message: expecting a JsString, got $invalidType")
-        case None => JsError(s"Missing @type field on a webSocket inbound message")
-      }
-    case _ => JsError("Expecting a JsObject to represent a webSocket inbound message")
-  }
-
-  private implicit val typeStateMapWrites: Writes[Map[TypeName, State]] = mapWrites[TypeName, State](_.asString(), stateWrites)
-  private implicit val typeStateWrites: Writes[TypeState] = Json.valueWrites[TypeState]
-  private implicit val changeWrites: OWrites[Map[AccountId, TypeState]] = mapWrites[AccountId, TypeState](_.id.value, typeStateWrites)
-  private implicit val pushStateWrites: Writes[PushState] = Json.valueWrites[PushState]
-  private implicit val stateChangeWrites: Writes[StateChange] = stateChange =>
-    stateChange.pushState.map(pushState =>
-      JsObject(Map(
-        "@type" -> JsString("StateChange"),
-        "changed" -> changeWrites.writes(stateChange.changes),
-        "pushState" -> pushStateWrites.writes(pushState))))
-      .getOrElse(
-        JsObject(Map(
-          "@type" -> JsString("StateChange"),
-          "changed" -> changeWrites.writes(stateChange.changes))))
-  private val stateChangeWritesNoPushState: Writes[StateChange] = stateChange =>
-      JsObject(Map(
-        "@type" -> JsString("StateChange"),
-        "changed" -> changeWrites.writes(stateChange.changes)))
-
-  private implicit val webSocketResponseWrites: Writes[WebSocketResponse] = response => {
-    val apiResponseJson: JsObject = responseObjectFormat.writes(response.responseObject)
-    JsObject(Map(
-      "@type" -> JsString("Response"),
-      "requestId" -> response.requestId.map(_.value).map(JsString).getOrElse(JsNull))
-      ++ apiResponseJson.value)
-  }
-  private implicit val webSocketErrorWrites: Writes[WebSocketError] = error => {
-    val errorJson: JsObject = problemDetailsWrites.writes(error.problemDetails)
-    JsObject(Map(
-      "@type" -> JsString("RequestError"),
-      "requestId" -> error.requestId.map(_.value).map(JsString).getOrElse(JsNull))
-      ++ errorJson.value)
-  }
-  private implicit val pingWrites: Writes[PingMessage] = Json.writes[PingMessage]
-  private implicit val webSocketOutboundWrites: Writes[OutboundMessage] = {
-    case pingMessage: PingMessage => pingWrites.writes(pingMessage)
-    case stateChange: StateChange => stateChangeWrites.writes(stateChange)
-    case response: WebSocketResponse => webSocketResponseWrites.writes(response)
-    case error: WebSocketError => webSocketErrorWrites.writes(error)
-  }
-  private val sseOutboundWrites: Writes[OutboundMessage] = {
-    case pingMessage: PingMessage => pingWrites.writes(pingMessage)
-    case stateChange: StateChange => stateChangeWritesNoPushState.writes(stateChange)
-    case response: WebSocketResponse => webSocketResponseWrites.writes(response)
-    case error: WebSocketError => webSocketErrorWrites.writes(error)
-  }
 
   def serialize(session: Session): JsValue = Json.toJson(session)
 
   def serialize(requestObject: RequestObject): JsValue = Json.toJson(requestObject)
 
-  def serialize(responseObject: ResponseObject): JsValue = Json.toJson(responseObject)
+  def serialize(responseObject: ResponseObject): JsObject = Json.toJsObject(responseObject)
 
-  def serialize(problemDetails: ProblemDetails): JsValue = Json.toJson(problemDetails)
+  def serialize(problemDetails: ProblemDetails): JsObject = Json.toJsObject(problemDetails)
 
   def serialize(errors: JsError): JsValue = Json.toJson(errors)
 
-  def serialize(outboundMessage: OutboundMessage): JsValue = Json.toJson(outboundMessage)
-
-  def serializeSSE(outboundMessage: OutboundMessage): JsValue = sseOutboundWrites.writes(outboundMessage)
-
   def deserializeRequestObject(input: String): JsResult[RequestObject] = Json.parse(input).validate[RequestObject]
 
-  def deserializeWebSocketInboundMessage(input: String): JsResult[WebSocketInboundMessage] = Try(Json.parse(input).validate[WebSocketInboundMessage])
-    .fold(e => JsError(e.getMessage), result => result)
-
   def deserializeRequestObject(input: InputStream): JsResult[RequestObject] = Json.parse(input).validate[RequestObject]
+
+  def deserializeRequestObject(js: JsValue): JsResult[RequestObject] = js.validate[RequestObject]
 
   def deserializeResponseObject(input: String): JsResult[ResponseObject] = Json.parse(input).validate[ResponseObject]
 }
